Key-Phrase-Extraction/requirements.txt
azure-ai-textanalytics==5.1.0